feat: user에게 값 입력 받는 로직 구현
