+++ src/main/java/racing/CarRacing.java
@@ -1,16 +1,5 @@
 package racing;
 
-import static racing.InputView.inputValue;
-
 public class CarRacing {
-    public static void main(String[] args) {
-
-        inputValue();
-
-    }
-
-//    private static forward(){
-//
-//    }
 
 }
