+++ src/main/java/racing/view/InputView.java
@@ -1,0 +1,18 @@
+package racing.view;
+
+import java.util.Scanner;
+
+public class InputView {
+
+    public static int inputCarCount() {
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("자동차 대수는 몇 대 인가요?");
+        return scanner.nextInt();
+    }
+
+    public static int inputGameCount() {
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("시도할 회수는 몇 회 인가요?");
+        return scanner.nextInt();
+    }
+}
