+++ src/main/java/racing/view/ResultView.java
@@ -1,0 +1,7 @@
+package racing.view;
+
+public class ResultView {
+    public static void showResult() {
+        System.out.println("실행 결과");
+    }
+}
