+++ src/test/java/study/CarRacingTest.java
@@ -1,5 +1,10 @@
 package study;
 
-public class CarRacingTest {
+import org.junit.jupiter.api.Test;
 
+public class CarRacingTest {
+    @Test
+    void name() {
+
+    }
 }
